Tests/LibWeb/Text/input/canvas/webgl-many-image-source-uploads.html

<!doctype html>
<script src="../include.js"></script>
<script src="webgl-helpers.js"></script>
<script>
    test(() => {
        const canvas = document.createElement("canvas");
        const gl = canvas.getContext("webgl");
        const source = new ImageData(new Uint8ClampedArray([1, 2, 3, 255]), 1, 1);
        const uploadCount = 130;

        for (let i = 0; i < uploadCount; ++i) {
            const texture = gl.createTexture();
            gl.bindTexture(gl.TEXTURE_2D, texture);
            gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
            gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
            gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, source);
        }

        println(`uploads: ${uploadCount}`);
        printError("error", gl);
    });
</script>
