Tests/LibWeb/Text/expected/canvas/webgl-many-image-source-uploads.txt

uploads: 130
error: NO_ERROR
